+++ packages/api/test/elastic/pages.test.ts
@@ -3,8 +3,6 @@ import { expect } from 'chai'
 import { InFilter, ReadFilter } from '../../src/utils/search'
 import {
   ArticleSavingRequestStatus,
-  Highlight,
-  Label,
   Page,
   PageContext,
   PageType,
@@ -21,62 +19,15 @@ import {
   searchPages,
   updatePage,
 } from '../../src/elastic/pages'
-import { addLabelInPage, setLabelsForHighlight } from '../../src/elastic/labels'
-import {
-  addHighlightToPage,
-  searchHighlights,
-} from '../../src/elastic/highlights'
+import { createTestElasticPage } from '../util'
 
-describe('elastic api', () => {
+describe('pages in elastic', () => {
   const userId = 'userId'
   const ctx: PageContext = {
     pubsub: createPubSubClient(),
     refresh: true,
     uid: userId,
   }
-
-  let page: Page
-
-  before(async () => {
-    // create a testing page
-    page = {
-      id: '',
-      hash: 'test hash',
-      userId: userId,
-      pageType: PageType.Article,
-      title: 'test title',
-      content: '<p>test</p>',
-      slug: 'test slug',
-      createdAt: new Date(),
-      updatedAt: new Date(),
-      savedAt: new Date(),
-      readingProgressPercent: 100,
-      readingProgressAnchorIndex: 0,
-      url: 'https://blog.omnivore.app/p/getting-started-with-omnivore',
-      archivedAt: new Date(),
-      labels: [
-        {
-          id: 'Test label id',
-          name: 'test label',
-          color: '#ffffff',
-          createdAt: new Date(),
-        },
-        {
-          id: 'Test label id 2',
-          name: 'test label 2',
-          color: '#eeeeee',
-          createdAt: new Date(),
-        },
-      ],
-      state: ArticleSavingRequestStatus.Succeeded,
-    }
-    page.id = (await createPage(page, ctx))!
-  })
-
-  after(async () => {
-    // delete the testing page
-    await deletePage(page.id, ctx)
-  })
 
   describe('createPage', () => {
     let newPageId: string
@@ -108,6 +59,18 @@ describe('elastic api', () => {
   })
 
   describe('getPageByParam', () => {
+    let page: Page
+
+    before(async () => {
+      // create a testing page
+      page = await createTestElasticPage(userId)
+    })
+
+    after(async () => {
+      // delete the testing page
+      await deletePage(page.id, ctx)
+    })
+
     it('gets a page by url', async () => {
       const pageFound = await getPageByParam({
         userId: page.userId,
@@ -119,6 +82,18 @@ describe('elastic api', () => {
   })
 
   describe('getPageById', () => {
+    let page: Page
+
+    before(async () => {
+      // create a testing page
+      page = await createTestElasticPage(userId)
+    })
+
+    after(async () => {
+      // delete the testing page
+      await deletePage(page.id, ctx)
+    })
+
     it('gets a page by id', async () => {
       const pageFound = await getPageById(page.id)
       expect(pageFound).not.undefined
@@ -126,6 +101,18 @@ describe('elastic api', () => {
   })
 
   describe('updatePage', () => {
+    let page: Page
+
+    before(async () => {
+      // create a testing page
+      page = await createTestElasticPage(userId)
+    })
+
+    after(async () => {
+      // delete the testing page
+      await deletePage(page.id, ctx)
+    })
+
     it('updates a page', async () => {
       const newTitle = 'new title'
       const updatedPageData: Partial<Page> = {
@@ -140,6 +127,18 @@ describe('elastic api', () => {
   })
 
   describe('searchPages', () => {
+    let page: Page
+
+    before(async () => {
+      // create a testing page
+      page = await createTestElasticPage(userId)
+    })
+
+    after(async () => {
+      // delete the testing page
+      await deletePage(page.id, ctx)
+    })
+
     it('searches pages', async () => {
       const searchResults = await searchPages(
         {
@@ -148,7 +147,7 @@ describe('elastic api', () => {
           inFilter: InFilter.ALL,
           labelFilters: [],
           readFilter: ReadFilter.ALL,
-          query: 'test',
+          query: page.content,
         },
         page.userId
       )
@@ -156,43 +155,12 @@ describe('elastic api', () => {
     })
   })
 
-  describe('addLabelInPage', () => {
-    context('when the label not exist in the page', () => {
-      it('adds the label to the page', async () => {
-        const newLabel = {
-          id: 'new label id',
-          name: 'new label',
-          color: '#07D2D1',
-        }
-
-        const result = await addLabelInPage(page.id, newLabel, ctx)
-        expect(result).to.be.true
-
-        const updatedPage = await getPageById(page.id)
-        expect(updatedPage?.labels).to.deep.include(newLabel)
-      })
-    })
-
-    context('when the label exists in the page', () => {
-      it('does not add the label to the page', async () => {
-        const newLabel = {
-          id: 'Test label id',
-          name: 'test label',
-          color: '#07D2D1',
-        }
-
-        const result = await addLabelInPage(page.id, newLabel, ctx)
-
-        expect(result).to.be.false
-      })
-    })
-  })
-
   describe('countByCreatedAt', () => {
     const createdAt = Date.now() - 60 * 60 * 24 * 1000
+    let page: Page
 
     before(async () => {
-      const newPageData: Page = {
+      page = {
         id: '',
         hash: 'hash',
         userId: userId,
@@ -208,93 +176,16 @@ describe('elastic api', () => {
         state: ArticleSavingRequestStatus.Succeeded,
       }
 
-      await createPage(newPageData, ctx)
-    })
-
-    it('counts pages by createdAt', async () => {
-      const count = await countByCreatedAt(userId, createdAt, createdAt)
-      expect(count).to.eq(1)
-    })
-  })
-
-  describe('searchHighlights', () => {
-    const highlightId = 'highlightId'
-
-    before(async () => {
-      const highlightData: Highlight = {
-        patch: 'test patch',
-        quote: 'test content',
-        shortId: 'test shortId',
-        id: highlightId,
-        userId: page.userId,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      }
-
-      await addHighlightToPage(page.id, highlightData, ctx)
-    })
-
-    it('searches highlights', async () => {
-      const [searchResults, count] = (await searchHighlights(
-        {
-          query: 'test',
-        },
-        page.userId
-      )) || [[], 0]
-
-      expect(count).to.eq(1)
-      expect(searchResults[0].id).to.eq(highlightId)
-    })
-  })
-
-  describe('setLabelsForHighlight', () => {
-    const page = {
-      id: 'testPageId',
-      hash: 'test set labels for highlight hash',
-      userId: userId,
-      pageType: PageType.Article,
-      title: 'test set labels for highlight title',
-      content: '<p>test</p>',
-      slug: 'test set labels for highlight slug',
-      createdAt: new Date(),
-      savedAt: new Date(),
-      readingProgressPercent: 100,
-      readingProgressAnchorIndex: 0,
-      url: 'https://blog.omnivore.app/p/setting-labels-for-highlight',
-      state: ArticleSavingRequestStatus.Succeeded,
-    }
-    const highlightId = 'highlightId'
-    const label: Label = {
-      id: 'test label id',
-      name: 'test label',
-      color: '#07D2D1',
-    }
-
-    before(async () => {
-      // create a testing page
-      await createPage(page, ctx)
-
-      const highlightData: Highlight = {
-        patch: 'test set labels patch',
-        quote: 'test set labels quote',
-        shortId: 'test set labels shortId',
-        id: highlightId,
-        userId: page.userId,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-      }
-
-      await addHighlightToPage(page.id, highlightData, ctx)
+      page.id = (await createPage(page, ctx))!
     })
 
     after(async () => {
       await deletePage(page.id, ctx)
     })
 
-    it('sets labels for highlights', async () => {
-      const result = await setLabelsForHighlight(highlightId, [label], ctx)
-
-      expect(result).to.be.true
+    it('counts pages by createdAt', async () => {
+      const count = await countByCreatedAt(userId, createdAt, createdAt)
+      expect(count).to.eq(1)
     })
   })
