+++ packages/api/test/util.ts
@@ -4,9 +4,8 @@ import { v4 } from 'uuid'
 import { corsConfig } from '../src/utils/corsConfig'
 import { ArticleSavingRequestStatus, Label, Page } from '../src/elastic/types'
 import { PageType } from '../src/generated/graphql'
-import { User } from '../src/entity/user'
 import { createPubSubClient } from '../src/datalayer/pubsub'
-import { createPage, getPageById } from '../src/elastic/pages'
+import { createPage } from '../src/elastic/pages'
 
 const { app, apollo } = createApp()
 export const request = supertest(app)
@@ -39,13 +38,13 @@ export const generateFakeUuid = () => {
 }
 
 export const createTestElasticPage = async (
-  user: User,
+  userId: string,
   labels?: Label[]
 ): Promise<Page> => {
   const page: Page = {
     id: '',
     hash: 'test hash',
-    userId: user.id,
+    userId,
     pageType: PageType.Article,
     title: 'test title',
     content: '<p>test content</p>',
@@ -59,19 +58,10 @@ export const createTestElasticPage = async (
     state: ArticleSavingRequestStatus.Succeeded,
   }
 
-  const pageId = await createPage(page, {
+  page.id = (await createPage(page, {
     pubsub: createPubSubClient(),
     refresh: true,
-    uid: user.id,
-  })
-  if (pageId) {
-    page.id = pageId
-  }
-
-  const res = await getPageById(page.id)
-  console.log('got page', res)
-  if (!res) {
-    throw new Error('Failed to create page')
-  }
-  return res
+    uid: userId,
+  }))!
+  return page
 }
